+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -11,6 +11,9 @@
 
 - Feature (`ngx-calendar`): Added new `@Input` named `showStartTimeInputs` and `showEndTimeInputs` set to a `boolean`
 - Feature (`ngx-column`): Added a new column component
+- Enhancement (`ngx-filter-select`): Added support to render an icon
+- Enhancement (`ngx-filter-select`): Added support to display the options popup at the right of the select
+- Enhancement (`ngx-filter-select`): Added support to select the SelectAll option automatically when all the options are selected
 
 ## 49.2.1 (2025-06-18)