+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -11,9 +11,11 @@
 
 - Feature (`ngx-calendar`): Added new `@Input` named `showStartTimeInputs` and `showEndTimeInputs` set to a `boolean`
 - Feature (`ngx-column`): Added a new column component
-- Enhancement (`ngx-filter-select`): Added support to render an icon
-- Enhancement (`ngx-filter-select`): Added support to display the options popup at the right of the select
+- Enhancement (`ngx-filter-select`): Added support to render an icon and its position
+- Enhancement (`ngx-filter-select`): Added support to display the options popup at the right or left of the select
 - Enhancement (`ngx-filter-select`): Added support to select the SelectAll option automatically when all the options are selected
+- Enhancement (`ngx-filter-select`): Added support to set a filter type (button and dropdown)
+- Added a new output to emit an event when the filter is clicked
 
 ## 49.2.1 (2025-06-18)